Denavit e Jacobiano

+++ Matlab/PC.m
@@ -6,7 +6,7 @@ d06 = matrizEntrada(1:end-1,4);
 R06 = matrizEntrada(1:3,1:3);
 
 k = [0;0;1];
-PC = d06 - (d6 * R06 * k);
+PC = double(d06 - (d6 * R06 * k));
 
 end
